changed temp.py into vsiualize_data.py and into the correct directory
